+++ frontend/src/components/CommentList.js
@@ -1,30 +1,6 @@
-import Owner from './Owner';
-import CreatedAt from './CreatedAt';
-import marked from 'marked';
-import DOMPurify from 'dompurify';
-import { makeStyles } from '@mui/styles';
-import Grid from '@mui/material/Grid';
+import Comment from './Comment';
 
 const CommentList = (props) => {
-  const styles = makeStyles({
-    comment: {
-      marginBottom: 30,
-    },
-    commentRight: {
-      border: 'solid 1px #bbb',
-      borderRadius: '10px',
-      overflowWrap: 'break-word',
-      padding: 10,
-    },
-    commentRightFooter: {
-      display: 'flex',
-      justifyContent: 'flex-end',
-      marginTop: 10,
-    },
-  });
-
-  const classes = styles();
-
   return (
     <div>
       {props.commentsAndUsers.length ? (
@@ -33,30 +9,9 @@ const CommentList = (props) => {
         <h2>コメントはまだありません</h2>
       )}
       {props.commentsAndUsers.map((commentAndUser, i) => (
-        <Grid container key={i} className={classes.comment}>
-          <Grid item xs={2} lg={1}>
-            <div style={{ marginTop: 10 }}>
-              <Owner
-                userId={commentAndUser.user.id}
-                userName={commentAndUser.user.name}
-                userIconUrl={commentAndUser.user.icon_url}
-              />
-            </div>
-          </Grid>
-          <Grid item xs={10} lg={11} className={classes.commentRight}>
-            <div
-              dangerouslySetInnerHTML={{
-                __html: DOMPurify.sanitize(
-                  marked(commentAndUser.comment.content)
-                ),
-              }}
-              style={{ borderBottom: 'solid 1px #bbb' }}
-            ></div>
-            <div className={classes.commentRightFooter}>
-              <CreatedAt createdAt={commentAndUser.comment.created_at} />
-            </div>
-          </Grid>
-        </Grid>
+        <div key={i}>
+          <Comment commentAndUser={commentAndUser} />
+        </div>
       ))}
     </div>
   );
